+++ CHANGELOG.md
@@ -10,6 +10,7 @@ To get the diff for a specific change, go to https://github.com/superdesk/web-pu
  * improvement [#630] Add option to exclude articles from content list items
  
 ### 1.6.0
+ * bug [#784] Fixed search pagination
  * bug [#763] Expose more metadata in article media
  * bug [#754] Always cast publishedBefore and publishedAfter parameters to object of type DateTime
  * feature [#758] Get next and previous article from content list (in template)